requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.3.1
Brotli==1.0.9
cffi==1.14.6
cssselect2==0.4.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
Django==3.1.6
django-3-jet==1.0.8
django-import-export==2.5.0
django-jet==1.0.8
django-jet-reboot==1.1.1
et-xmlfile==1.1.0
filelock==3.0.12
fonttools==4.27.0
future==0.18.2
google-api-python-client==1.4.1
html5lib==1.1
httplib2==0.19.1
MarkupPy==1.14
mysqlclient==2.0.3
oauth2client==4.1.3
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.2.0
pisa==3.0.33
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydyf==0.1.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyphen==0.11.0
python-bidi==0.4.2
pytz==2021.1
PyYAML==5.4.1
reportlab==3.6.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
tinycss2==1.1.0
uritemplate==3.0.1
virtualenv==20.4.2
virtualenvwrapper-win==1.2.6
weasyprint==53.3
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==2.0.1
XlsxWriter==3.0.1
xlwt==1.3.0
zopfli==0.1.8
